+++ R/template_arguments.R
@@ -413,6 +413,15 @@ template_arguments <- function(
           paste0(path, "/", tfound[i]))
       }
       
+      # Read provenance -------------------------------------------------------
+      
+      if (stringr::str_detect(
+        tfound[i], 
+        attr_tmp$regexpr[attr_tmp$template_name == "provenance"])) {
+        templates[[i]]$content <- read_tbl(
+          paste0(path, "/", tfound[i]))
+      }
+      
       # Read taxonomic coverage -----------------------------------------------
       
       if (stringr::str_detect(